pyproject.toml
[project]
name = "inventory-analysitc"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.62.0",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
]
